Add a channel for queuing tweets, process tweets off that channel, better comments and some names for functions, go fmt
