+++ helpers.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"errors"
+	"github.com/dghubble/go-twitter/twitter"
 	"go.uber.org/zap"
 	"net/http"
 )
@@ -18,4 +19,24 @@ func extractID(w http.ResponseWriter, r *http.Request) (string, error) {
 	log.Infow("Getting record", zap.String("id", id[0]))
 
 	return id[0], nil
+}
+
+// getInReplyToTwitterUser is a helper function takes the screen name of a Twitter user (IE - the honorary on an invoked
+// tweet) and returns the full Twitter user details for that user
+func getInReplyToTwitterUser(sn string) *twitter.User {
+	// in the event of a retweet with comment where a user is invoking Charity Yeti, there will be no screenname so we
+	// should return early
+	if sn == "" {
+		log.Error("No screen name provided and cannot get honorary user details")
+		return nil
+	}
+	user, _, err := twitterClient.Users.Show(&twitter.UserShowParams{
+		ScreenName: sn,
+	})
+
+	if err != nil {
+		log.Errorf("cannot get honorary user details: %v", err)
+	}
+
+	return user
 }
